requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Automat==20.2.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cycler==0.11.0
EasyModbus==1.4.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
kiwisolver==1.3.2
matplotlib==3.4.3
MinimalModbus==0.5
multidict==5.1.0
numpy==1.21.4
pandas==1.3.4
Pillow==8.4.0
prompt-toolkit==3.0.20
Pygments==2.10.0
pymodbus==2.5.2
pyparsing==3.0.6
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyserial==3.5
pyserial-asyncio==0.5
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing-extensions==3.10.0.2
wcwidth==0.2.5
yarl==1.6.3
zope.interface==5.4.0
python-dotenv==0.19.2
